third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/webrtc/RTCRtpTransceiver-stop-expected.txt

This is a testharness.js-based test.
PASS A transceiver added and stopped before the initial offer generation should not trigger an offer m-section generation
PASS During renegotiation, adding and stopping a transceiver should not trigger a renegotiated offer m-section generation
PASS A stopped sendonly transceiver should generate an inactive m-section in the offer
PASS A stopped inactive transceiver should generate an inactive m-section in the offer
PASS If a transceiver is stopped locally, setting a locally generated answer should still work
PASS If a transceiver is stopped remotely, setting a locally generated answer should still work
PASS If a transceiver is stopped, transceivers, senders and receivers should disappear after offer/answer
FAIL If a transceiver is stopped, transceivers should end up in state stopped assert_equals: expected "recvonly" but got "stopped"
Harness: the test ran to completion.
